fixed LazyLoadingFriendsControllerMock conflicts
